Use PaymentFlow's extension customer's email resolving
